requirement.txt
aiohttp==3.9.1
asyncio
pytest==8.3.4
pytest-asyncio==0.23.1
python-dotenv==1.0.1
